[#10]refactor: 함수 이동, 인덴트 맞춤
