+++ app/route/services/ai_service.py
@@ -5,17 +5,17 @@ from time import sleep
 import aiofiles
 from openai import OpenAIError, APITimeoutError, AsyncOpenAI
 
-from app._config.settings import Settings
+from app.config.settings import Settings
 from app.route.models.correct_response import CorrectResponse, ModifiedCode
 from app.route.models.hint_response import HintResponse
 from app.route.services.exception.enum.error_enum import ErrorEnum
 from app.route.services.exception.openai_exception import OpenaiException
-from app.route.services.prompts.prompt_file import PromptFile
+from app.route.services.prompts.prompt_file_name import PromptFileName
 from app.web.logger import logger
 
 
 async def correct(code: str) -> CorrectResponse:
-    template_path = path.join(path.dirname(__file__), 'prompts', PromptFile.CORRECT_TEMPLATE)
+    template_path = path.join(path.dirname(__file__), 'prompts', PromptFileName.CORRECT_TEMPLATE)
     template = await _load_template(template_path)
 
     prompt = template.format(code=code)
@@ -25,7 +25,7 @@ async def correct(code: str) -> CorrectResponse:
 
 
 async def hint(line: int, code: str) -> HintResponse:
-    template_path = path.join(path.dirname(__file__), 'prompts', PromptFile.HINT_TEMPLATE)
+    template_path = path.join(path.dirname(__file__), 'prompts', PromptFileName.HINT_TEMPLATE)
     template = await _load_template(template_path)
 
     prompt = template.format(line=line, code=code)
@@ -37,16 +37,6 @@ async def hint(line: int, code: str) -> HintResponse:
 async def _load_template(file_path) -> str:
     async with aiofiles.open(file_path, 'r') as file:
         return await file.read()
-
-
-async def _parse_correct_response(response_data: dict) -> CorrectResponse:
-    modified_codes_data = response_data["modified_codes"]
-    # ModifiedCode 리스트 생성
-    modified_codes = [
-        ModifiedCode(line=item["line"], code=item["code"])
-        for item in modified_codes_data
-    ]
-    return CorrectResponse(reason=response_data["reason"], modified_codes=modified_codes)
 
 
 async def _call_openai_api(prompt: str,  max_retries: int = 2, delay: int = 1) -> dict:
@@ -75,18 +65,9 @@ async def _call_openai_api(prompt: str,  max_retries: int = 2, delay: int = 1) -
                 logger.error(f"OpenAI API 연결 실패: {e}. {delay}초 후 재시도 ({retries}/{max_retries})")
                 sleep(delay)
             else:
-                raise OpenaiException(ErrorEnum.OPNEAI_SERVER_ERROR, _extract_openai_error(e))
+                raise OpenaiException(ErrorEnum.OPNEAI_SERVER_ERROR, e)
 
         except OpenAIError as e:
-            raise OpenaiException(ErrorEnum.OPNEAI_SERVER_ERROR, _extract_openai_error(e))
+            raise OpenaiException(ErrorEnum.OPNEAI_SERVER_ERROR, e)
 
 
-def _extract_openai_error(e: OpenAIError) -> dict:
-    response = getattr(e, 'response', None)  # response가 없으면 None 반환
-    status_code = response.status_code if response else "Unknown"  # response가 없으면 'Unknown'
-
-    return {
-        "error_type": type(e).__name__,  # 예외 클래스 이름
-        "message": str(e),  # 예외 메시지
-        "status_code": status_code  # 상태 코드
-    }
