+++ app/route/services/prompts/prompt_file_name.py
@@ -1,4 +1,3 @@
-class PromptFile:
-
+class PromptFileName:
     CORRECT_TEMPLATE = 'correct_template.txt'
     HINT_TEMPLATE = 'hint_template.txt'
