+++ app/web/exception_handlers.py
@@ -1,3 +1,4 @@
+from openai import OpenAIError
 from starlette import status
 from fastapi import FastAPI, Request
 from fastapi.responses import JSONResponse
@@ -26,7 +27,9 @@ def setup_exception_handlers(app: FastAPI):
     @app.exception_handler(OpenaiException)
     async def openai_exception_handler(request: Request, exc: OpenaiException):
         logger.info(
-            f"{exc.error_enum.code} - OpenAI exception occurred for URI: {request.url}. Error: {exc.result}. ")
+            f"{exc.error_enum.code} - OpenAI exception occurred for URI: {request.url}.\n"
+            f"          Error: {_get_unique_value_in_openAI_error(exc.result)}. "
+        )
 
         response = ErrorResponse(
             code=ErrorEnum.TASK_FAIL.code,
@@ -61,3 +64,13 @@ def setup_exception_handlers(app: FastAPI):
             status_code=status.HTTP_400_BAD_REQUEST,
             content=response.to_dict()
         )
+
+    def _get_unique_value_in_openAI_error(e: OpenAIError) -> dict:
+        response = getattr(e, 'response', None)  # response가 없으면 None 반환
+        status_code = response.status_code if response else "Unknown"  # response가 없으면 'Unknown'
+
+        return {
+            "error_type": type(e).__name__,  # 예외 클래스 이름
+            "message": str(e),  # 예외 메시지
+            "status_code": status_code  # 상태 코드
+        }
